Move functions createMQConnector and getSpecializedMQConnector to Utilities
